# openapi.yml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Mailejoe API
  description: The Mailejoe REST API
  license:
    name: MIT
servers:
  - url: https://app.mailejoe.com/api/v1
paths:
  '/setup':
    post:
      summary: Sets up a new organization
      operationId: setupOrganization
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetupOrganizationBody'
      responses:
        204:
          description: Successfully created a new organization
        400:
          description: The payload body is malformed or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: An internal server error has occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    SetupOrganizationBody:
      type: object
      properties:
        name:
          type: string
          description: The name of the organization to create
        firstName:
          type: string
          description: The first name of the administrator of the organization
        lastName:
          type: string
          description: The last name of the administrator of the organization
        email:
          type: string
          description: The email of the administrator of the organization
      required:
        - name
        - firstName
        - lastName
        - email
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: The error message from the API
      required:
        - error